+++ pkg/workflow/compiler.go
@@ -1692,26 +1692,7 @@ func (c *Compiler) buildMainJob(data *WorkflowData, jobName string, taskJobCreat
 		needsRoleCheck := c.needsRoleCheck(data, frontmatter)
 
 		if needsRoleCheck {
-			// Add team member check step
-			steps = append(steps, "      - name: Check team membership for workflow\n")
-			steps = append(steps, "        id: check-team-member\n")
-			steps = append(steps, "        uses: actions/github-script@v8\n")
-
-			// Add environment variables for permission check
-			steps = append(steps, "        env:\n")
-			steps = append(steps, fmt.Sprintf("          GITHUB_AW_REQUIRED_ROLES: %s\n", strings.Join(data.Roles, ",")))
-
-			steps = append(steps, "        with:\n")
-			steps = append(steps, "          script: |\n")
-
-			// Generate the JavaScript code for the role check
-			scriptContent := c.generateRoleCheckScript(data.Roles)
-			scriptLines := strings.Split(scriptContent, "\n")
-			for _, line := range scriptLines {
-				if strings.TrimSpace(line) != "" {
-					steps = append(steps, fmt.Sprintf("            %s\n", line))
-				}
-			}
+			steps = c.generateRoleCheck(data, steps)
 		}
 	}
 
@@ -2048,167 +2029,9 @@ func (c *Compiler) generatePrompt(yaml *strings.Builder, data *WorkflowData) {
 	}
 
 	// Add cache folder notification if cache-memory is enabled
-	if data.CacheMemoryConfig != nil && data.CacheMemoryConfig.Enabled {
-		yaml.WriteString("          \n")
-		yaml.WriteString("          ---\n")
-		yaml.WriteString("          \n")
-		yaml.WriteString("          ## Cache Folder Available\n")
-		yaml.WriteString("          \n")
-		yaml.WriteString("          You have access to a persistent cache folder at `/tmp/cache-memory/` where you can read and write files to create memories and store information.\n")
-		yaml.WriteString("          \n")
-		yaml.WriteString("          - **Read/Write Access**: You can freely read from and write to any files in this folder\n")
-		yaml.WriteString("          - **Persistence**: Files in this folder persist across workflow runs via GitHub Actions cache\n")
-		yaml.WriteString("          - **Last Write Wins**: If multiple processes write to the same file, the last write will be preserved\n")
-		yaml.WriteString("          - **File Share**: Use this as a simple file share - organize files as you see fit\n")
-		yaml.WriteString("          \n")
-		yaml.WriteString("          Examples of what you can store:\n")
-		yaml.WriteString("          - `/tmp/cache-memory/notes.txt` - general notes and observations\n")
-		yaml.WriteString("          - `/tmp/cache-memory/preferences.json` - user preferences and settings\n")
-		yaml.WriteString("          - `/tmp/cache-memory/history.log` - activity history and logs\n")
-		yaml.WriteString("          - `/tmp/cache-memory/state/` - organized state files in subdirectories\n")
-		yaml.WriteString("          \n")
-		yaml.WriteString("          Feel free to create, read, update, and organize files in this folder as needed for your tasks.\n")
-	}
+	generateCacheMemoryPromptSection(yaml, data.CacheMemoryConfig)
 
-	if data.SafeOutputs != nil {
-		// Add output instructions for all engines (GITHUB_AW_SAFE_OUTPUTS functionality)
-		yaml.WriteString("          \n")
-		yaml.WriteString("          ---\n")
-		yaml.WriteString("          \n")
-		yaml.WriteString("          ## ")
-		written := false
-		if data.SafeOutputs.AddComments != nil {
-			yaml.WriteString("Adding a Comment to an Issue or Pull Request")
-			written = true
-		}
-		if data.SafeOutputs.CreateIssues != nil {
-			if written {
-				yaml.WriteString(", ")
-			}
-			yaml.WriteString("Creating an Issue")
-		}
-		if data.SafeOutputs.CreatePullRequests != nil {
-			if written {
-				yaml.WriteString(", ")
-			}
-			yaml.WriteString("Creating a Pull Request")
-		}
-
-		if data.SafeOutputs.AddLabels != nil {
-			if written {
-				yaml.WriteString(", ")
-			}
-			yaml.WriteString("Adding Labels to Issues or Pull Requests")
-			written = true
-		}
-
-		if data.SafeOutputs.UpdateIssues != nil {
-			if written {
-				yaml.WriteString(", ")
-			}
-			yaml.WriteString("Updating Issues")
-			written = true
-		}
-
-		if data.SafeOutputs.PushToPullRequestBranch != nil {
-			if written {
-				yaml.WriteString(", ")
-			}
-			yaml.WriteString("Pushing Changes to Branch")
-			written = true
-		}
-
-		if data.SafeOutputs.CreateCodeScanningAlerts != nil {
-			if written {
-				yaml.WriteString(", ")
-			}
-			yaml.WriteString("Creating Code Scanning Alert")
-			written = true
-		}
-
-		// Missing-tool is always available
-		if written {
-			yaml.WriteString(", ")
-		}
-		yaml.WriteString("Reporting Missing Tools or Functionality")
-
-		yaml.WriteString("\n")
-		yaml.WriteString("          \n")
-		yaml.WriteString("          **IMPORTANT**: To do the actions mentioned in the header of this section, use the **safe-outputs** tools, do NOT attempt to use `gh`, do NOT attempt to use the GitHub API. You don't have write access to the GitHub repo.\n")
-		yaml.WriteString("          \n")
-
-		if data.SafeOutputs.AddComments != nil {
-			yaml.WriteString("          **Adding a Comment to an Issue or Pull Request**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To add a comment to an issue or pull request, use the add-comments tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-
-		if data.SafeOutputs.CreateIssues != nil {
-			yaml.WriteString("          **Creating an Issue**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To create an issue, use the create-issue tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-
-		if data.SafeOutputs.CreatePullRequests != nil {
-			yaml.WriteString("          **Creating a Pull Request**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To create a pull request:\n")
-			yaml.WriteString("          1. Make any file changes directly in the working directory\n")
-			yaml.WriteString("          2. If you haven't done so already, create a local branch using an appropriate unique name\n")
-			yaml.WriteString("          3. Add and commit your changes to the branch. Be careful to add exactly the files you intend, and check there are no extra files left un-added. Check you haven't deleted or changed any files you didn't intend to.\n")
-			yaml.WriteString("          4. Do not push your changes. That will be done by the tool.\n")
-			yaml.WriteString("          5. Create the pull request with the create-pull-request tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-
-		if data.SafeOutputs.AddLabels != nil {
-			yaml.WriteString("          **Adding Labels to Issues or Pull Requests**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To add labels to an issue or a pull request, use the add-labels tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-
-		if data.SafeOutputs.UpdateIssues != nil {
-			yaml.WriteString("          **Updating an Issue**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To udpate an issue, use the update-issue tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-
-		if data.SafeOutputs.PushToPullRequestBranch != nil {
-			yaml.WriteString("          **Pushing Changes to Pull Request Branch**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To push changes to the branch of a pull request:\n")
-			yaml.WriteString("          1. Make any file changes directly in the working directory\n")
-			yaml.WriteString("          2. Add and commit your changes to the local copy of the pull request branch. Be careful to add exactly the files you intend, and check there are no extra files left un-added. Check you haven't deleted or changed any files you didn't intend to.\n")
-			yaml.WriteString("          3. Push the branch to the repo by using the push-to-pr-branch tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-
-		if data.SafeOutputs.CreateCodeScanningAlerts != nil {
-			yaml.WriteString("          **Creating Code Scanning Alert**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To create code scanning alert use the create-code-scanning-alert tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-
-		// Missing-tool instructions are only included when configured
-		if data.SafeOutputs.MissingTool != nil {
-			yaml.WriteString("          **Reporting Missing Tools or Functionality**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To report a missing tool use the missing-tool tool from the safe-outputs MCP.\n")
-			yaml.WriteString("          \n")
-		}
-
-		if data.SafeOutputs.CreatePullRequestReviewComments != nil {
-			yaml.WriteString("          **Creating a Pull Request Review Comment**\n")
-			yaml.WriteString("          \n")
-			yaml.WriteString("          To create a pull request review comment, use the create-pull-request-review-comment tool from the safe-outputs MCP\n")
-			yaml.WriteString("          \n")
-		}
-	}
+	generateSafeOutputsPromptSection(yaml, data.SafeOutputs)
 
 	yaml.WriteString("          EOF\n")
 
@@ -2250,253 +2073,6 @@ func (c *Compiler) extractJobsFromFrontmatter(frontmatter map[string]any) map[st
 		}
 	}
 	return make(map[string]any)
-}
-
-// extractSafeOutputsConfig extracts output configuration from frontmatter
-func (c *Compiler) extractSafeOutputsConfig(frontmatter map[string]any) *SafeOutputsConfig {
-	var config *SafeOutputsConfig
-
-	if output, exists := frontmatter["safe-outputs"]; exists {
-		if outputMap, ok := output.(map[string]any); ok {
-			config = &SafeOutputsConfig{}
-
-			// Handle create-issue
-			issuesConfig := c.parseIssuesConfig(outputMap)
-			if issuesConfig != nil {
-				config.CreateIssues = issuesConfig
-			}
-
-			// Handle create-discussion
-			discussionsConfig := c.parseDiscussionsConfig(outputMap)
-			if discussionsConfig != nil {
-				config.CreateDiscussions = discussionsConfig
-			}
-
-			// Handle add-comment
-			commentsConfig := c.parseCommentsConfig(outputMap)
-			if commentsConfig != nil {
-				config.AddComments = commentsConfig
-			}
-
-			// Handle create-pull-request
-			pullRequestsConfig := c.parsePullRequestsConfig(outputMap)
-			if pullRequestsConfig != nil {
-				config.CreatePullRequests = pullRequestsConfig
-			}
-
-			// Handle create-pull-request-review-comment
-			prReviewCommentsConfig := c.parsePullRequestReviewCommentsConfig(outputMap)
-			if prReviewCommentsConfig != nil {
-				config.CreatePullRequestReviewComments = prReviewCommentsConfig
-			}
-
-			// Handle create-code-scanning-alert
-			securityReportsConfig := c.parseCodeScanningAlertsConfig(outputMap)
-			if securityReportsConfig != nil {
-				config.CreateCodeScanningAlerts = securityReportsConfig
-			}
-
-			// Parse allowed-domains configuration
-			if allowedDomains, exists := outputMap["allowed-domains"]; exists {
-				if domainsArray, ok := allowedDomains.([]any); ok {
-					var domainStrings []string
-					for _, domain := range domainsArray {
-						if domainStr, ok := domain.(string); ok {
-							domainStrings = append(domainStrings, domainStr)
-						}
-					}
-					config.AllowedDomains = domainStrings
-				}
-			}
-
-			// Parse add-labels configuration
-			if labels, exists := outputMap["add-labels"]; exists {
-				if labelsMap, ok := labels.(map[string]any); ok {
-					labelConfig := &AddLabelsConfig{}
-
-					// Parse allowed labels (optional)
-					if allowed, exists := labelsMap["allowed"]; exists {
-						if allowedArray, ok := allowed.([]any); ok {
-							var allowedStrings []string
-							for _, label := range allowedArray {
-								if labelStr, ok := label.(string); ok {
-									allowedStrings = append(allowedStrings, labelStr)
-								}
-							}
-							labelConfig.Allowed = allowedStrings
-						}
-					}
-
-					// Parse max (optional)
-					if maxCount, exists := labelsMap["max"]; exists {
-						// Handle different numeric types that YAML parsers might return
-						var maxCountInt int
-						var validMaxCount bool
-						switch v := maxCount.(type) {
-						case int:
-							maxCountInt = v
-							validMaxCount = true
-						case int64:
-							maxCountInt = int(v)
-							validMaxCount = true
-						case uint64:
-							maxCountInt = int(v)
-							validMaxCount = true
-						case float64:
-							maxCountInt = int(v)
-							validMaxCount = true
-						}
-						if validMaxCount {
-							labelConfig.MaxCount = &maxCountInt
-						}
-					}
-
-					// Parse github-token
-					if githubToken, exists := labelsMap["github-token"]; exists {
-						if githubTokenStr, ok := githubToken.(string); ok {
-							labelConfig.GitHubToken = githubTokenStr
-						}
-					}
-
-					config.AddLabels = labelConfig
-				} else if labels == nil {
-					// Handle null case: create empty config (allows any labels)
-					config.AddLabels = &AddLabelsConfig{}
-				}
-			}
-
-			// Handle update-issue
-			updateIssuesConfig := c.parseUpdateIssuesConfig(outputMap)
-			if updateIssuesConfig != nil {
-				config.UpdateIssues = updateIssuesConfig
-			}
-
-			// Handle push-to-pr-branch
-			pushToBranchConfig := c.parsePushToPullRequestBranchConfig(outputMap)
-			if pushToBranchConfig != nil {
-				config.PushToPullRequestBranch = pushToBranchConfig
-			}
-
-			// Handle missing-tool (parse configuration if present)
-			missingToolConfig := c.parseMissingToolConfig(outputMap)
-			if missingToolConfig != nil {
-				config.MissingTool = missingToolConfig
-			}
-
-			// Handle staged flag
-			if staged, exists := outputMap["staged"]; exists {
-				if stagedBool, ok := staged.(bool); ok {
-					config.Staged = &stagedBool
-				}
-			}
-
-			// Handle env configuration
-			if env, exists := outputMap["env"]; exists {
-				if envMap, ok := env.(map[string]any); ok {
-					config.Env = make(map[string]string)
-					for key, value := range envMap {
-						if valueStr, ok := value.(string); ok {
-							config.Env[key] = valueStr
-						}
-					}
-				}
-			}
-
-			// Handle github-token configuration
-			if githubToken, exists := outputMap["github-token"]; exists {
-				if githubTokenStr, ok := githubToken.(string); ok {
-					config.GitHubToken = githubTokenStr
-				}
-			}
-
-			// Handle max-patch-size configuration
-			if maxPatchSize, exists := outputMap["max-patch-size"]; exists {
-				switch v := maxPatchSize.(type) {
-				case int:
-					if v >= 1 {
-						config.MaximumPatchSize = v
-					}
-				case int64:
-					if v >= 1 {
-						config.MaximumPatchSize = int(v)
-					}
-				case uint64:
-					if v >= 1 {
-						config.MaximumPatchSize = int(v)
-					}
-				case float64:
-					intVal := int(v)
-					if intVal >= 1 {
-						config.MaximumPatchSize = intVal
-					}
-				}
-			}
-
-			// Set default value if not specified or invalid
-			if config.MaximumPatchSize == 0 {
-				config.MaximumPatchSize = 1024 // Default to 1MB = 1024 KB
-			}
-		}
-	}
-
-	return config
-}
-
-// extractCacheMemoryConfig extracts cache-memory configuration from tools section
-func (c *Compiler) extractCacheMemoryConfig(tools map[string]any) *CacheMemoryConfig {
-	cacheMemoryValue, exists := tools["cache-memory"]
-	if !exists {
-		return nil
-	}
-
-	config := &CacheMemoryConfig{}
-
-	// Handle boolean value (simple enable/disable)
-	if boolValue, ok := cacheMemoryValue.(bool); ok {
-		config.Enabled = boolValue
-		if config.Enabled {
-			// Set defaults
-			config.Key = "memory-${{ github.workflow }}-${{ github.run_id }}"
-		}
-		return config
-	}
-
-	// Handle object configuration
-	if configMap, ok := cacheMemoryValue.(map[string]any); ok {
-		config.Enabled = true
-
-		// Set defaults
-		config.Key = "memory-${{ github.workflow }}-${{ github.run_id }}"
-
-		// Parse custom key
-		if key, exists := configMap["key"]; exists {
-			if keyStr, ok := key.(string); ok {
-				config.Key = keyStr
-				// Automatically append -${{ github.run_id }} if the key doesn't already end with it
-				runIdSuffix := "-${{ github.run_id }}"
-				if !strings.HasSuffix(config.Key, runIdSuffix) {
-					config.Key = config.Key + runIdSuffix
-				}
-			}
-		}
-
-		// Parse retention days
-		if retentionDays, exists := configMap["retention-days"]; exists {
-			if retentionDaysInt, ok := retentionDays.(int); ok {
-				config.RetentionDays = &retentionDaysInt
-			} else if retentionDaysFloat, ok := retentionDays.(float64); ok {
-				retentionDaysIntValue := int(retentionDaysFloat)
-				config.RetentionDays = &retentionDaysIntValue
-			} else if retentionDaysUint64, ok := retentionDays.(uint64); ok {
-				retentionDaysIntValue := int(retentionDaysUint64)
-				config.RetentionDays = &retentionDaysIntValue
-			}
-		}
-
-		return config
-	}
-
-	return nil
 }
 
 // buildCustomJobs creates custom jobs defined in the frontmatter jobs section
